# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sous Mon Toit",
        "description": "Documentation for the API SousMonToit",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Server of the real estate agency Sous Mon Toit"
        },
        {
            "url": "https://sousmontoit.herokuapp.com/",
            "description": "Server heroku of the real estate agency Sous Mon Toit"
        }
    ],
    "paths": {
...
        "/contract/": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Get list of all contracts",
                "description": "Return list of all contracts",
                "operationId": "getContractsList",
                "responses": {
                    "200": {
                        "description": "A list with contracts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Contracts"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "default": {
                        "$ref": "#/components/responses/default"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/contract/{id}": {
            "get": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Get contract information",
                "description": "Return one contract",
                "operationId": "getContract",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contracts"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "default": {
                        "$ref": "#/components/responses/default"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/contract/saveContract": {
            "post": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Store new contract",
                "description": "Returns successful message",
                "operationId": "storeContract",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contracts"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contracts"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "default": {
                        "$ref": "#/components/responses/default"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/contract/update/{id}": {
            "put": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Update existing contract",
                "description": "Returns updated contract data",
                "operationId": "updateContract",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Contracts"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Contracts"
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "default": {
                        "$ref": "#/components/responses/default"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/contract/archive/{id}": {
            "delete": {
                "tags": [
                    "Contracts"
                ],
                "summary": "Delete existing contract",
                "description": "Return succesfull message",
                "operationId": "archiveContract",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "403": {
                        "$ref": "#/components/responses/403"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "default": {
                        "$ref": "#/components/responses/default"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Contracts": {
                "title": "Contracts",
                "description": "Contracts model",
                "properties": {
                    "id": {
                        "description": "ID of the user"
                    },
                    "created_at": {
                        "description": "Date of created"
                    },
                    "updated_at": {
                        "description": "Date of updated"
                    },
                    "deleted_at": {
                        "description": "Date of archived"
                    },
                    "folder": {
                        "description": "Folder where the contract is stocked"
                    },
                    "name": {
                        "description": "Name of the contract"
                    },
                    "id_staff": {
                        "description": "Id of the staff"
                    },
                    "id_estate": {
                        "description": "Id of the estate"
                    },
                    "id_customer": {
                        "description": "Id of the customer"
                    },
                    "id_contract_type": {
                        "description": "Id of the contract type"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "400": {
                "description": "Bad Request"
            },
            "401": {
                "description": "Access token is missing or invalid"
            },
            "403": {
                "description": "Forbidden"
            },
            "404": {
                "description": "Resource Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Resource Not Found"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "default": {
                "description": "une erreur \"inattendue\""
            }
        },
        "parameters": {
            "id": {
                "name": "id",
                "in": "path",
                "description": "Id of resource",
                "required": true,
                "schema": {
                    "type": "integer"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter the token for customer or staff",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}